stripe-cloud-function/requirements.txt
datetime
functions-framework
google-cloud
google-cloud-bigquery
stripe